+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.6.0 (unreleased)
+
+### 💥 Breaking changes
+
+* Drop support of doctrine/dbal < 4
+* Drop support of doctrine/orm < 3
+
+### 🐛 Bug fixes
+
+* Fix Doctrine DBAL exception when adding table to schema
+* Fix the `--adapter` option of the `webf:flysystem-failover:list-messages` command, which was not filtering anything
+
 ## v0.5.1 (February 8, 2025)
 
 ### ✨ New features